+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - Fix scraper is looping over failing URLs forever (#379)
+- Fix extraction of header content for stackoverflow domains (#378)
 
 ## [3.0.0] - 2025-12-12